feat(banner): added beta banner
