+++ frontend/src/components/layout/BetaBanner.tsx
@@ -1,29 +1,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
 import '@/components/layout/BetaBanner.scss';
-import { useState, useEffect } from 'react';
-
-const useMediaQuery = (query: string) => {
-  const getMatches = () =>
-    typeof window !== 'undefined' && window.matchMedia
-      ? window.matchMedia(query).matches
-      : false;
-
-  const [matches, setMatches] = useState(getMatches);
-
-  useEffect(() => {
-    if (typeof window === 'undefined' || !window.matchMedia) return;
-
-    const media = window.matchMedia(query);
-    const listener = (e: MediaQueryListEvent) => setMatches(e.matches);
-
-    setMatches(media.matches);
-    media.addEventListener('change', listener);
-    return () => media.removeEventListener('change', listener);
-  }, [query]);
-
-  return matches;
-};
+import useMediaQuery from '@/hooks/useMediaQuery';
 
 const BetaBanner = () => {
   const isMobile = useMediaQuery('(max-width: 767px)');
